Move ldap to own file and add some logging
